Refactor stackHorizontal method into a new module
- Cut the stackHorizontal method from the Registry class and added it to
  a new module for better organization and reusability.
